+++ lib/processor/filter.go
@@ -40,7 +40,7 @@ Tests each message against a condition, if the condition fails then the message
 is dropped. You can find a [full list of conditions here](../conditions).
 
 NOTE: If you are combining messages into batches using the
-` + "[`combine`](#combine) or [`batch`](#batch)" + ` processors this filter will
+` + "[`batch`](#batch)" + ` processor this filter will
 apply to the _whole_ batch. If you instead wish to filter _individual_ parts of
 the batch use the ` + "[`filter_parts`](#filter_parts)" + ` processor.`,
 		sanitiseConfigFunc: func(conf Config) (interface{}, error) {